first commit
There's a server which manages games, and a client which shows the
games.

The person who starts a game is in charge and can see everything.

The next two people to join are the players. They each see only their
boards.

Anyone else who joins is an observer and can see everything.

There is a built-in list of mathematicians taken from
mathigon.org/timeline. I haven't included the images in this commit
because I'm not sure about the licensing.